Review of the way we get the `www.` version and vice versa of an element to test.
Indeed, we now generate them with the `domains.list` in order to have PyFunceble testing them on the fly.
It for sure increase the time of each tests but it is the right thing to do as we cannot have a domain
inaccessible from `www.xxx.com` but accessible from `xxx.com` and vice-versa.